# pet-metric-server/docs/openapi.yaml
openapi: 3.1.0
info:
  title: PetMetric Content API
  version: 1.0.0
  description: Read-only REST API for the PetMetric public website.
servers:
  - url: http://localhost:4000/api/v1
    description: Local development
tags:
  - name: Home
  - name: Categories
  - name: Products
  - name: Comparisons
  - name: Best picks
  - name: Guides
  - name: Troubleshooting
  - name: Pages
paths:
  /home:
    get:
      tags: [Home]
      operationId: getHome
      summary: Get all homepage content
      responses:
        '200': { $ref: '#/components/responses/Success' }
  /categories:
    get:
      tags: [Categories]
      operationId: listCategories
      summary: List published categories
      responses:
        '200': { $ref: '#/components/responses/Success' }
  /categories/{slug}:
    get:
      tags: [Categories]
      operationId: getCategory
      summary: Get one category
      parameters: [{ $ref: '#/components/parameters/Slug' }]
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '404': { $ref: '#/components/responses/NotFound' }
  /categories/{slug}/products:
    get:
      tags: [Categories, Products]
      operationId: listCategoryProducts
      summary: List products in one category
      parameters:
        - { $ref: '#/components/parameters/Slug' }
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/Limit' }
        - { $ref: '#/components/parameters/Search' }
        - { $ref: '#/components/parameters/Sort' }
        - { $ref: '#/components/parameters/MaxPrice' }
        - { $ref: '#/components/parameters/CatCount' }
        - { $ref: '#/components/parameters/LitterType' }
        - { $ref: '#/components/parameters/AppConnectivity' }
      responses:
        '200': { $ref: '#/components/responses/ListSuccess' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
  /products:
    get:
      tags: [Products]
      operationId: listProducts
      summary: List and filter products
      parameters:
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/Limit' }
        - { $ref: '#/components/parameters/Category' }
        - { $ref: '#/components/parameters/Search' }
        - { $ref: '#/components/parameters/Sort' }
        - { $ref: '#/components/parameters/MaxPrice' }
        - { $ref: '#/components/parameters/Featured' }
        - { $ref: '#/components/parameters/CatCount' }
        - { $ref: '#/components/parameters/LitterType' }
        - { $ref: '#/components/parameters/AppConnectivity' }
      responses:
        '200': { $ref: '#/components/responses/ListSuccess' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /products/{slug}:
    get:
      tags: [Products]
      operationId: getProduct
      summary: Get a product and its review
      parameters: [{ $ref: '#/components/parameters/Slug' }]
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '404': { $ref: '#/components/responses/NotFound' }
  /comparisons:
    get:
      tags: [Comparisons]
      operationId: listComparisons
      summary: List comparisons
      parameters: [{ $ref: '#/components/parameters/Page' }, { $ref: '#/components/parameters/Limit' }, { $ref: '#/components/parameters/Category' }, { $ref: '#/components/parameters/Sort' }, { $ref: '#/components/parameters/Featured' }]
      responses:
        '200': { $ref: '#/components/responses/ListSuccess' }
  /comparisons/{slug}:
    get:
      tags: [Comparisons]
      operationId: getComparison
      summary: Get a comparison with criteria
      parameters: [{ $ref: '#/components/parameters/Slug' }]
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '404': { $ref: '#/components/responses/NotFound' }
  /best-picks:
    get:
      tags: [Best picks]
      operationId: listBestPicks
      summary: List scenario-based best picks
      parameters: [{ $ref: '#/components/parameters/Page' }, { $ref: '#/components/parameters/Limit' }, { $ref: '#/components/parameters/Category' }]
      responses:
        '200': { $ref: '#/components/responses/ListSuccess' }
  /best-picks/{slug}:
    get:
      tags: [Best picks]
      operationId: getBestPick
      summary: Get one best pick
      parameters: [{ $ref: '#/components/parameters/Slug' }]
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '404': { $ref: '#/components/responses/NotFound' }
  /guides:
    get:
      tags: [Guides]
      operationId: listGuides
      summary: List buying and educational guides
      parameters: [{ $ref: '#/components/parameters/Page' }, { $ref: '#/components/parameters/Limit' }, { $ref: '#/components/parameters/Category' }, { $ref: '#/components/parameters/Search' }]
      responses:
        '200': { $ref: '#/components/responses/ListSuccess' }
  /guides/{slug}:
    get:
      tags: [Guides]
      operationId: getGuide
      summary: Get one guide
      parameters: [{ $ref: '#/components/parameters/Slug' }]
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '404': { $ref: '#/components/responses/NotFound' }
  /troubleshooting:
    get:
      tags: [Troubleshooting]
      operationId: listTroubleshooting
      summary: List troubleshooting guides with steps
      parameters: [{ $ref: '#/components/parameters/Page' }, { $ref: '#/components/parameters/Limit' }, { $ref: '#/components/parameters/Category' }, { $ref: '#/components/parameters/Search' }, { $ref: '#/components/parameters/Sort' }]
      responses:
        '200': { $ref: '#/components/responses/ListSuccess' }
  /troubleshooting/{slug}:
    get:
      tags: [Troubleshooting]
      operationId: getTroubleshooting
      summary: Get one troubleshooting guide
      parameters: [{ $ref: '#/components/parameters/Slug' }]
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '404': { $ref: '#/components/responses/NotFound' }
  /pages/{slug}:
    get:
      tags: [Pages]
      operationId: getPage
      summary: Get a generic published page
      parameters: [{ $ref: '#/components/parameters/Slug' }]
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '404': { $ref: '#/components/responses/NotFound' }
components:
  parameters:
    Slug: { name: slug, in: path, required: true, schema: { type: string, pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*$' } }
    Page: { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
    Limit: { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
    Category: { name: category, in: query, schema: { type: string } }
    Search: { name: search, in: query, schema: { type: string, maxLength: 100 } }
    Sort: { name: sort, in: query, schema: { type: string, enum: [rank, price_asc, price_desc, score_desc, popular, newest], default: rank } }
    MaxPrice: { name: maxPrice, in: query, schema: { type: number, minimum: 0 } }
    CatCount: { name: catCount, in: query, schema: { type: string } }
    LitterType: { name: litterType, in: query, schema: { type: string } }
    AppConnectivity: { name: app, in: query, schema: { type: string } }
    Featured: { name: featured, in: query, schema: { type: boolean } }
  schemas:
    SuccessEnvelope:
      type: object
      required: [data]
      properties: { data: {} }
    ListEnvelope:
      type: object
      required: [data, meta]
      properties:
        data: { type: array, items: {} }
        meta:
          type: object
          required: [page, limit, total, totalPages]
          properties:
            page: { type: integer }
            limit: { type: integer }
            total: { type: integer }
            totalPages: { type: integer }
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties:
            code: { type: string }
            message: { type: string }
            details: {}
            requestId: { type: string }
  responses:
    Success:
      description: Success
      content: { application/json: { schema: { $ref: '#/components/schemas/SuccessEnvelope' } } }
    ListSuccess:
      description: Paginated success
      content: { application/json: { schema: { $ref: '#/components/schemas/ListEnvelope' } } }
    BadRequest:
      description: Invalid parameters
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
    NotFound:
      description: Resource not found
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }